backend/requirements/base.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
dj-database-url==2.3.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
Pillow>=11.2.1,<13
whitenoise==6.8.2
gunicorn==23.0.0
